Relocated train/predict data processing scripts, updated imports and paths
